refactor(docs): move common logic of most popular pages into utils
